Use size_t

+++ sfsim2025/list.c
@@ -11,7 +11,7 @@ list_t *make_list(void) {
   return result;
 }
 
-static void grow_list(list_t *list, int element_size, bool atomic) {
+static void grow_list(list_t *list, size_t element_size, bool atomic) {
   if (list->buffer_size < (list->size + 1) * element_size) {
     list->buffer_size = list->buffer_size ? 2 * list->buffer_size : element_size;
     GLuint *space = atomic ? GC_MALLOC_ATOMIC(list->buffer_size) : GC_MALLOC(list->buffer_size);
